Add a helper class to read in area names in regions.
