+++ EU4ToVic2Tests/EU4ToVic2Tests.vcxproj
@@ -168,6 +168,7 @@
     <ClCompile Include="..\EU4toV2\Source\Configuration.cpp" />
     <ClCompile Include="..\EU4toV2\Source\EU4World\EU4Version.cpp" />
     <ClCompile Include="..\EU4toV2\Source\EU4World\Regions\Area.cpp" />
+    <ClCompile Include="..\EU4toV2\Source\EU4World\Regions\AreaNames.cpp" />
     <ClCompile Include="..\EU4toV2\Source\EU4World\Regions\Areas.cpp" />
     <ClCompile Include="..\EU4toV2\Source\V2World\BlockedTechSchools.cpp" />
     <ClCompile Include="..\EU4toV2\Source\V2World\V2TechSchools.cpp" />
@@ -177,6 +178,7 @@
     <ClCompile Include="..\googletest\googletest\src\gtest-all.cc" />
     <ClCompile Include="..\googletest\googletest\src\gtest_main.cc" />
     <ClCompile Include="ConfigurationTests.cpp" />
+    <ClCompile Include="EU4ToVic2Tests\AreaNamesTests.cpp" />
     <ClCompile Include="EU4WorldTests\AreasTests.cpp" />
     <ClCompile Include="EU4WorldTests\EU4AreaTests.cpp" />
     <ClCompile Include="EU4WorldTests\EU4VersionTests.cpp" />
