+++ EU4ToVic2Tests/EU4ToVic2Tests.vcxproj.filters
@@ -73,6 +73,12 @@
     <ClCompile Include="EU4WorldTests\AreasTests.cpp">
       <Filter>EU4WorldTests</Filter>
     </ClCompile>
+    <ClCompile Include="EU4ToVic2Tests\AreaNamesTests.cpp">
+      <Filter>EU4WorldTests</Filter>
+    </ClCompile>
+    <ClCompile Include="..\EU4toV2\Source\EU4World\Regions\AreaNames.cpp">
+      <Filter>EU4ToVic2Files\EU4World\Regions</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <Filter Include="EU4ToVic2Files">
