+++ EU4toV2/Source/EU4World/Regions/AreaNames.cpp
@@ -1,4 +1,4 @@
-/*Copyright (c) 2018 The Paradox Game Converters Project
+/*Copyright (c) 2019 The Paradox Game Converters Project
 
 Permission is hereby granted, free of charge, to any person obtaining
 a copy of this software and associated documentation files (the
@@ -21,58 +21,15 @@ SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.*/
 
 
 
-#ifndef EU4_REGIONS_H
-#define EU4_REGIONS_H
+#include "AreaNames.h"
 
 
 
-#include "newParser.h"
-#include "Regions/Region.h"
-#include <map>
-#include <memory>
-#include <set>
-#include <string>
-
-
-
-class Object;
-namespace EU4
+EU4::AreaNames::AreaNames(std::istream& theStream)
 {
-	class areas;
+	registerKeyword(std::regex("[a-zA-Z0-9_]+"), [this](const std::string & areaName, std::istream & theStream) {
+		names.insert(areaName);
+	});
 
-	class Regions: commonItems::parser
-	{
-		public:
-			static bool provinceInRegion(int province, const std::string& regionName)
-			{
-				return getInstance()->ProvinceInRegion(province, regionName);
-			}
-
-		private:
-			static Regions* instance;
-			static Regions* getInstance()
-			{
-				if (instance == nullptr)
-				{
-					instance = new Regions;
-				}
-				return instance;
-			}
-
-			Regions();
-
-			void initEU4RegionsOldVersion();
-
-			void initEU4RegionsNewVersion();
-			void initEU4RegionsFile(const EU4::areas& areas, const std::string& regionsFilename);
-
-			bool ProvinceInRegion(int province, const std::string& regionName);
-
-			std::map<int, std::set<std::string>> EU4RegionsMap;
-			std::map<std::string, EU4::region> regions;
-	};
+	parseStream(theStream);
 }
-
-
-
-#endif //EU4_REGIONS_H
